Rename and focus for bg atlas 2008
